Better tests, updated docs to match current status of library
