+++ README.rst
@@ -129,6 +129,15 @@ Sample output::
     Include dirs: ['foo/bar', 'spam/eggs']
     Compiling: ('x.cpp', 'y.cpp', 'z.cpp')
 
+**Color controls** ::
+
+    from plumbum import COLOR
+    with COLOR.RED:
+        print("This library provides safe, flexible color access.")
+        print("Color", "(and styles in general)" << COLOR.BOLD, "are easy!")
+    print("The simple 8 colors or", 'All 265' << COLOR['Orchid'] + COLOR.UNDERLINE, 'can be used.')
+    print("Unsafe " + COLOR.BG['DarkKhaki'] + "color access" - COLOR.BG + " is available too.")
+
 
 
 .. image:: https://d2weczhvl823v0.cloudfront.net/tomerfiliba/plumbum/trend.png
